Use JobService with AsyncTask for calendar sync
