+++ app/src/main/AndroidManifest.xml
@@ -327,6 +327,11 @@
             android:name=".quest.widgets.AgendaWidgetService"
             android:permission="android.permission.BIND_REMOTEVIEWS" />
 
+        <service
+            android:name=".app.sync.AndroidCalendarSyncJobService"
+            android:permission="android.permission.BIND_JOB_SERVICE"
+            android:exported="true"/>
+
         <meta-data
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
